cms/services/cms/pyproject.toml
[tool.poetry]
name = "cms"
version = "1.0.0"
description = "Content Management System built with FastAPI"
authors = ["Mahmoud Elshaffie <mahmoud.elshaffie@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = {extras = ["email"], version = "^2.11.7"}
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
emails = "^0.6"
jinja2 = "^3.1.0"
aiofiles = "^24.1.0"
pillow = "^11.3.0"
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
requests = "^2.32.5"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.17.1"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true